refactor: Refactor SnackBar and move SearchHeader component to shared package
