+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/createDukan/content/CreateDukanContent.kt
@@ -71,7 +71,8 @@ private fun CreateDukanScaffold(
             state.snackBarState?.let { snackBarState ->
                 SnackBar(
                     snackBarUiState = snackBarState,
-                    onDismiss = listener::onDismissSnackBar
+                    onDismiss = listener::onDismissSnackBar,
+                    onClick = listener::onDismissSnackBar
                 )
             }
         },
