+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/createProduct/CreateProductScreen.kt
@@ -187,11 +187,9 @@ private fun CreateProductSnackBar(
 ) {
     state.snackBarUiState?.let { snackBarState ->
         SnackBar(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable(onClick = interactionListener::onDismissSnackBar),
+            snackBarUiState = snackBarState,
             onDismiss = interactionListener::onDismissSnackBar,
-            snackBarUiState = snackBarState
+            onClick = interactionListener::onDismissSnackBar
         )
     }
 }
