+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/dukanCart/content/DukanCartContent.kt
@@ -54,7 +54,8 @@ fun DukanCartContent(
             state.snackBarState?.let { snackBarState ->
                 SnackBar(
                     snackBarUiState = snackBarState,
-                    onDismiss = listener::onDismissSnackBar
+                    onDismiss = listener::onDismissSnackBar,
+                    onClick = listener::onDismissSnackBar
                 )
             }
         }
