+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/editProduct/EditProductScreen.kt
@@ -1,7 +1,6 @@
 package net.thechance.mena.dukan.presentation.screen.editProduct
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -186,11 +185,9 @@ private fun EditProductSnackBar(
 ) {
     state.snackBarUiState?.let { snackBarState ->
         SnackBar(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable(onClick = interactionListener::onDismissSnackBar),
+            snackBarUiState = snackBarState,
             onDismiss = interactionListener::onDismissSnackBar,
-            snackBarUiState = snackBarState
+            onClick = interactionListener::onDismissSnackBar
         )
     }
 }
