+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/manageDukan/component/ManageDukanSnackBar.kt
@@ -27,7 +27,8 @@ fun ManageDukanSnackbar(
             state.snackBarState?.let {
                 SnackBar(
                     snackBarUiState = state.snackBarState,
-                    onDismiss = listener::onDismissSnackBar
+                    onDismiss = listener::onDismissSnackBar,
+                    onClick = listener::onDismissSnackBar
                 )
             }
         }
