+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/manageShelf/content/ManageShelfContent.kt
@@ -56,7 +56,8 @@ fun ManageShelfContent(
             state.snackBarState?.let { snackBarState ->
                 SnackBar(
                     snackBarUiState = snackBarState,
-                    onDismiss = listener::onDismissSnackBar
+                    onDismiss = listener::onDismissSnackBar,
+                    onClick = listener::onDismissSnackBar
                 )
             }
         }
