+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/productDetails/ProductDetailsContent.kt
@@ -56,7 +56,8 @@ fun ProductDetailsContent(
             state.snackBarState?.let { snackBarUiState ->
                 SnackBar(
                     snackBarUiState = snackBarUiState,
-                    onDismiss = listener::onDismissSnackBar
+                    onDismiss = listener::onDismissSnackBar,
+                    onClick = listener::onDismissSnackBar
                 )
             }
         }
