+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/shelfDetails/ShelfDetailsScreen.kt
@@ -87,7 +87,8 @@ private fun ShelfDetailsContent(
             state.snackBarState?.let { snackBarState ->
                 SnackBar(
                     snackBarUiState = snackBarState,
-                    onDismiss = listener::onDismissSnackBar
+                    onDismiss = listener::onDismissSnackBar,
+                    onClick = listener::onDismissSnackBar
                 )
             }
         }
